Split unittest type stubs to mirror unittest module.
