+++ stdlib/3/unittest/signals.pyi
@@ -1,0 +1,14 @@
+from typing import Any, Callable, overload, TypeVar
+import unittest.result
+
+
+_F = TypeVar('_F', bound=Callable[..., Any])
+
+
+def installHandler() -> None: ...
+def registerResult(result: unittest.result.TestResult) -> None: ...
+def removeResult(result: unittest.result.TestResult) -> bool: ...
+@overload
+def removeHandler() -> None: ...
+@overload
+def removeHandler(function: _F) -> _F: ...
