+++ stdlib/3/unittest/suite.pyi
@@ -1,0 +1,16 @@
+from typing import Iterable, Iterator, Union
+import unittest.case
+import unittest.result
+
+
+_TestType = Union[unittest.case.TestCase, TestSuite]
+
+
+class TestSuite(Iterable[_TestType]):
+    def __init__(self, tests: Iterable[_TestType] = ...) -> None: ...
+    def addTest(self, test: _TestType) -> None: ...
+    def addTests(self, tests: Iterable[_TestType]) -> None: ...
+    def run(self, result: unittest.result.TestResult) -> unittest.result.TestResult: ...
+    def debug(self) -> None: ...
+    def countTestCases(self) -> int: ...
+    def __iter__(self) -> Iterator[_TestType]: ...
